Linter configuration for golangci-lint added
This is motivated by #89 to document the target baseline for linting.
It intentionally avoids enforcing linting for changes prior to the
introduction of this configuration so that the project can be
incrementally brought in line.

Each linter is explicitly enabled so there are no surprises if the
defaults are changed.

Only simple formatters are included since the needs of this project
are relatively simple.